requirements.txt
python-telegram-bot==20.7
pyTelegramBotAPI==4.14.1
Flask==3.0.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.3.0
Werkzeug>=3.0.0
Jinja2==3.1.2 